apply method to packets WIP
